+++ web/angular-wallet/src/app/components/components.module.ts
@@ -13,33 +13,37 @@ import {RecipientInputComponent} from './recipient-input/recipient-input.compone
 import {AmountInputComponent} from './amount-input/amount-input.component';
 import {Ng5SliderModule} from 'ng5-slider';
 import {AppSharedModule} from '../shared/shared.module';
-import {FeeSelectorComponent} from './fee-selector/fee-selector.component';
 import {PageComponent} from './page/page.component';
 import {WarnSendDialogComponent} from './warn-send-dialog/warn-send-dialog.component';
 import {MatDialogModule} from '@angular/material/dialog';
 import { WatermarkComponent } from './watermark/watermark.component';
 import { LinkButtonComponent } from './link-button/link-button.component';
+import { FeeInputComponent } from './fee-input/fee-input.component';
+import { MessageInputComponent } from './message-input/message-input.component';
+import { MatCheckboxModule } from '@angular/material/checkbox';
 
 @NgModule({
   declarations: [
     SubmitTransactionComponent,
     RecipientInputComponent,
     AmountInputComponent,
-    FeeSelectorComponent,
     PageComponent,
     WarnSendDialogComponent,
     WatermarkComponent,
     LinkButtonComponent,
+    FeeInputComponent,
+    MessageInputComponent
   ],
   exports: [
     SubmitTransactionComponent,
     RecipientInputComponent,
     AmountInputComponent,
-    FeeSelectorComponent,
     PageComponent,
     WarnSendDialogComponent,
     WatermarkComponent,
-    LinkButtonComponent
+    LinkButtonComponent,
+    FeeInputComponent,
+    MessageInputComponent
   ],
   imports: [
     AppSharedModule,
@@ -57,6 +61,7 @@ import { LinkButtonComponent } from './link-button/link-button.component';
     Ng5SliderModule,
     RouterModule,
     MatDialogModule,
+    MatCheckboxModule
   ],
   entryComponents: [
     WarnSendDialogComponent