+++ web/angular-wallet/src/app/constants.ts
@@ -103,6 +103,15 @@ export const constants = {
       testnet: false
     },
     {
+      name: 'Signum Node Latin America',
+      region: 'Middle America',
+      location: 'Miami',
+      address: 'https://latam.signum.network',
+      port: 443,
+      reliable: true,
+      testnet: false
+    },
+    {
       name: 'Signum Node Canada',
       region: 'North America',
       location: 'Toronto',
@@ -130,6 +139,33 @@ export const constants = {
       testnet: false
     },
     {
+      name: 'Signum Node EU #2',
+      region: 'Europe',
+      location: 'Frankfurt',
+      address: 'https://europe2.signum.network',
+      port: 443,
+      reliable: true,
+      testnet: false
+    },
+    {
+      name: 'Signum Node US East',
+      region: 'United States',
+      location: 'North Virgina',
+      address: 'https://us-east.signum.network',
+      port: 443,
+      reliable: true,
+      testnet: false
+    },
+    {
+      name: 'Signum Node Singapore',
+      region: 'Singapore',
+      location: 'North Virgina',
+      address: 'https://singapore.signum.network',
+      port: 443,
+      reliable: true,
+      testnet: false
+    },
+    {
       name: 'Signum Node AU',
       region: 'Australia',
       location: 'Sydney',
@@ -148,19 +184,19 @@ export const constants = {
       testnet: false
     },
     {
-      name: 'Testnet Node Zoh',
+      name: 'Testnet Node Local',
       region: 'Europe',
-      location: 'Denmark',
-      address: 'https://testnet.burstcoin.network',
+      location: 'UK',
+      address: 'http://europe3.testnet.signum.network',
       port: 6876,
       reliable: false,
       testnet: true
     },
     {
-      name: 'Testnet Node Nivok',
-      region: 'Europe',
-      location: 'UK',
-      address: 'http://nivbox.co.uk',
+      name: 'Testnet Local',
+      region: 'None',
+      location: 'Local',
+      address: 'http://localhost',
       port: 6876,
       reliable: false,
       testnet: true